finish migration to activity extensions decouple lights dialog from its implementation
